fixed the constructor for fireboy and watergirl

+++ core/src/com/mygdx/game/Fireboy.java
@@ -22,7 +22,7 @@ public class Fireboy extends Character {
      * @param x a float representing the x position on the screen
      * @param y a float representing the y position on the screen
      */
-    public Fireboy(int x, int y) {
+    public Fireboy(float x, float y) {
         super(x, y);
     }
